requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
gunicorn==22.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg==3.1.18
python-dotenv==1.0.1
typing_extensions==4.11.0
Werkzeug==3.0.1
